requirements.txt
requests
bs4
numpy
pandas
lxml
pandasql 
webdriver-manager
geopandas
geoplot
matplotlib
fake-useragent
seaborn
python-dotenv
plotly
pyarrow
html5lib
squarify
dash
google
scrapy
google-cloud-bigquery
hsluv
googlemaps
wordcloud 
plotly-geo
geopandas
pyshp
shapely